Update HardFault handler to use asm to repair stack
After excercising the HardFault handler, it was obvious that stack usage
would not be possible if SP is out of bounds, because that could be
what caused the HardFault to begin with. The asm code is based on an
embedded blog post, and basically checks the SP against linker defined
limits (__StackTop and __StackLimit) to determine if overflow or not,
then uses the .asciz to pass the correct string to assert_failed.
I still need to verify that bcs is the appropriate conditional branch that
I want (and finish adding comments for each asm line).